+++ src/components/Detail/Detail.css
