+++ src/components/Detail/NoUserFound.js
@@ -1,6 +1,6 @@
 import React, { useEffect }from 'react';
 import { useHistory } from "react-router-dom";
-import noUser from '../assets/nouser.jpg';
+import noUser from '../../assets/nouser.jpg';
 
 import './NoUserFound.css';
