+++ src/components/RocketGrid/Grid.js
@@ -1,5 +1,5 @@
 import React from 'react';
-// import './RocketGrid.css';
+
 import './Grid.css';
 
 import HeaderCell from './HeaderCell';
